Move implementation class to impl package
