+++ src/test/java/org/folio/consortia/service/HttpRequestServiceImplTest.java
@@ -6,6 +6,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
 import org.folio.consortia.base.BaseUnitTest;
+import org.folio.consortia.service.impl.HttpRequestServiceImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
